backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0
